4.21 done. blogs can now only be removed if the user has the right token for that blog

+++ controllers/blogs.js
@@ -2,22 +2,15 @@ const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
+const logger = require('../utils/logger')
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog
-  .find({}).populate('user', {"username": 1, "name": 1, "id": 1})
+    .find({}).populate('user', { "username": 1, "name": 1, "id": 1 })
   response
     .json(blogs)
     .status(200)
 })
-
-// const getTokenFrom = request => {
-//   const authorization = request.get('authorization')
-//   if (authorization && authorization.startsWith('Bearer ')) {
-//     request.token = authorization.replace('Bearer ', '')
-//   }
-//   return null
-// }
 
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
@@ -45,18 +38,37 @@ blogsRouter.post('/', async (request, response) => {
 
 blogsRouter.delete('/:id', async (request, response, next) => {
   try {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET);
+    logger.info("decodedToken:", decodedToken);
+
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: 'token invalid' });
+    }
+
+    const blog = await Blog.findById(request.params.id);
+    logger.info("blog", blog);
+
+    if (blog.user.toString() === decodedToken.id.toString()) {
+      logger.info("blog.user === decodedtoken.id");
+
+      await Blog.findByIdAndRemove(request.params.id);
+      logger.info("blog removed");
+
+      response.status(204).end();
+    } else {
+      return response.status(401).json({ error: 'token invalid' });
+    }
   } catch (exception) {
-    next(exception)
+    logger.error("exception:", exception);
+    next(exception);
   }
-}
-)
+})
+
 
 blogsRouter.put('/:id', async (request, response, next) => {
   try {
     const body = request.body
-    
+
     const blog = {
       title: body.title,
       author: body.author,
@@ -65,8 +77,8 @@ blogsRouter.put('/:id', async (request, response, next) => {
     }
     const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
     response
-    .json(updatedBlog)
-    .status(200)
+      .json(updatedBlog)
+      .status(200)
 
   } catch (exception) {
     next(exception)
